requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.6
python-jose
flask-cors
six~=1.16.0
python-dotenv~=0.21.0
requests~=2.28.1
authlib~=1.1.0

pip~=22.0.4
toml~=0.10.2
cryptography~=38.0.3
rsa~=4.9
pyasn1~=0.4.8
ecdsa~=0.18.0
PyYAML~=6.0
MarkupSafe~=2.1.1
Werkzeug~=2.2.2
Jinja2~=3.0.3
click~=8.0.4
itsdangerous~=2.1.1
cachetools~=5.2.0
certifi~=2021.10.8
urllib3~=1.26.6
cffi~=1.15.1
google-api-python-client~=2.65.0
protobuf==3.20.1
libcst~=0.4.7
PyJWT~=2.3.0
Django~=4.0.3
idna~=3.3
pycparser~=2.21
setuptools~=57.4.0
future~=0.18.2
attrs~=22.1.0
Pillow~=8.3.1
sqlparse~=0.4.2
asgiref~=3.5.0
colorama~=0.4.4
numpy~=1.23.1
pyparsing~=2.4.7
httplib2~=0.21.0
uritemplate~=4.1.1
zipp~=3.8.1